requirements.txt
streamlit
# the pyavagen package fails with pillow versions 10 and above
pillow==9.5.0
prisma
faker
# pyavagen is used on windows only
pyavagen
# py-avataaars is used on linux/mac
py-avataaars
